+++ packages/garden/src/lib/components/DefaultGardenItem/index.ts
@@ -1,0 +1,1 @@
+export * from './DefaultGardenItem';
