+++ packages/garden/src/lib/components/FilterSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterSelector';
