+++ packages/garden/src/lib/components/GardenItemContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './GardenItemContainer';
