+++ packages/garden/src/lib/components/HeaderContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './HeaderContainer';
