+++ packages/garden/src/lib/components/SubGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './subGroup.styles';
