+++ packages/garden/src/lib/components/VirtualContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './VirtualContainer';
