+++ packages/garden/src/lib/components/VirtualGarden/index.ts
@@ -1,0 +1,1 @@
+export * from './VirtualGarden';
