+++ packages/garden/src/lib/hooks/index.ts
@@ -6,3 +6,4 @@ export * from './useGroupingKeys';
 export * from './useItemWidths';
 export * from './useRefresh';
 export * from './useVirtualScrolling';
+export * from './useGardenContext';
