+++ packages/garden/src/lib/hooks/useGardenContext.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { GardenContext } from '../components';
+
+export const useGardenContext = () => useContext(GardenContext);
